break out legend and fix avg heat load endpoint
