+++ heat-stack/app/components/ui/heat/CaseSummaryComponents/Graphs/HeatLoad.tsx
@@ -7,7 +7,6 @@ import {
 	CartesianGrid,
 	Tooltip,
 	ResponsiveContainer,
-	Legend,
 	Label,
 	Scatter,
 } from 'recharts'
@@ -16,7 +15,6 @@ import { Icon } from '../../../icon'
 import { HeatLoadGrid } from '../HeatLoadGrid'
 import {
 	COLOR_GREY_LIGHT,
-	COLOR_WHITE,
 	COLOR_ORANGE,
 	COLOR_BLUE,
 } from '../constants'
@@ -26,32 +24,7 @@ import {
 } from '../utility/heat-load-calculations'
 import { buildHeatLoadGraphData } from '../utility/build-heat-load-graph-data'
 import { HeatLoadGraphToolTip } from './HeatLoadGraphToolTip'
-
-// Custom legend component
-const CustomLegend = () => {
-  const legendItems = [
-    { name: "Maximum, no internal or solar gain", color: "#FFA500", type: "line" },
-    { name: "Average, with internal & solar gain", color: "#0000FF", type: "line" },
-    { name: "Maximum at design temperature", color: "#FFA500", type: "diamond" },
-    { name: "Average at design temperature", color: "#0000FF", type: "diamond" }
-  ];
-
-  return (
-    <div className="absolute top-6 right-6 bg-white border border-gray-200 rounded p-4 shadow-sm">
-      {legendItems.map((item, index) => (
-        <div key={index} className="flex items-center gap-2 mb-2 last:mb-0">
-          {item.type === 'line' ? (
-            <div className="w-8 h-0.5" style={{ backgroundColor: item.color }} />
-          ) : (
-            <div className="w-3 h-3 rotate-45" style={{ backgroundColor: item.color }} />
-          )}
-          <span className="text-sm">{item.name}</span>
-        </div>
-      ))}
-    </div>
-  );
-};
-
+import { CustomLegend } from './HeatLoadGraphLegend'
 
 const BUFFER_PERCENTAGE_MAX = 1.3; // 30% buffer
 const Y_AXIS_ROUNDING_UNIT = 10000; // Rounding unit for minY and maxY
